units WIP

+++ objects/Coordinate.h
@@ -9,9 +9,9 @@
 class Coordinate {
 
 public:
-    int x;
-    int y;
-    int z;
+    double x;
+    double y;
+    double z;
     Coordinate(): x(0), y(0), z(0){};
     Coordinate(int _x, int _y, int _z) : x(_x), y(_y), z(_z){}
